@@ -1,63 +1,50 @@
 {
-  "name": "MeleeGuessr",
-  "version": "1.3.0",
+  "name": "meleeguessr",
+  "version": "2.0.0",
   "private": true,
+  "author": "noahpodgurski",
   "dependencies": {
-    "@capacitor/android": "3.5.1",
-    "@capacitor/app": "1.1.1",
-    "@capacitor/core": "3.5.1",
-    "@capacitor/haptics": "1.1.4",
-    "@capacitor/keyboard": "1.2.2",
-    "@capacitor/status-bar": "1.0.8",
-    "@ionic/core": "^6.1.4",
-    "@ionic/react": "^6.1.4",
-    "axios": "^0.27.2",
-    "jwt-decode": "^3.1.2",
-    "mdb-react-ui-kit": "^3.0.0",
-    "mdbreact": "^5.2.0",
-    "react": "^18.0.0",
-    "react-bootstrap": "^2.3.1",
-    "react-dom": "^18.0.0",
-    "react-hot-toast": "^2.2.0",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
-    "recharts": "^2.1.9",
-    "typescript": "^4.6.3",
-    "web-vitals": "^2.1.4"
+    "@jsonjoy.com/json-pack": "^1.0.2",
+    "@shelacek/ubjson": "^1.1.1",
+    "@solid-primitives/raf": "^2.1.4",
+    "@solid-primitives/upload": "^0.0.105",
+    "@solidjs/router": "^0.14.3",
+    "@suid/material": "^0.17.0",
+    "@tanstack/solid-virtual": "^3.0.0-beta.6",
+    "@thisbeyond/solid-select": "^0.13.0",
+    "@zag-js/accordion": "^0.65.1",
+    "@zag-js/dialog": "^0.65.1",
+    "@zag-js/menu": "^0.65.1",
+    "@zag-js/solid": "^0.65.1",
+    "axios": "^1.7.4",
+    "fflate": "^0.7.4",
+    "hono": "^4.2.8",
+    "jwt-decode": "^4.0.0",
+    "solid": "^0.2.1",
+    "solid-js": "^1.8.21",
+    "solid-toast": "^0.5.0",
+    "web-vitals": "^4.2.3",
+    "zoo-ids": "^2.0.7"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build && npx cap copy",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "wrangler pages dev dist",
+    "types": "tsc",
+    "format": "prettier -w src/"
   },
   "devDependencies": {
-    "@capacitor/cli": "3.5.1",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.0.1",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^16.11.27",
-    "@types/react": "^18.0.5",
-    "@types/react-dom": "^18.0.1",
-    "sass": "^1.51.0"
+    "@cloudflare/workers-types": "^4.20240423.0",
+    "autoprefixer": "^10.4.13",
+    "postcss": "^8.4.19",
+    "prettier": "^2.8.1",
+    "prettier-plugin-tailwindcss": "^0.2.0",
+    "sass-embedded": "^1.77.8",
+    "tailwindcss": "^3.2.4",
+    "typescript": "^5.5.4",
+    "vite": "^5.4.2",
+    "vite-plugin-solid": "^2.10.2",
+    "vite-tsconfig-paths": "^5.0.1",
+    "wrangler": "^3.72.1"
   }
 }
