@@ -1,11 +1,9 @@
-import { MDBDataTableV5 } from "mdbreact";
-import React, { useContext, useState } from "react";
+import { Component, createEffect, createSignal } from 'solid-js';
 import './Table.scss';
 import UserService from "../services/user.service";
 import { GetStat } from "../models/Stat";
-import { useEffect } from "react";
-import { ILoading, LoadingContext } from "../hooks/UseLoader";
-import toast from "react-hot-toast";
+import { LoadingContext } from "../components/common/Loader";
+// import { createToast } from "../components/common/toaster";
 
 export type LeaderboardData = {
   user: string; 
@@ -13,15 +11,14 @@ export type LeaderboardData = {
 }
 
 
-const Leaderboards: React.FC = () => {
-  const [data, setData] = useState<LeaderboardData[]>([]);
-  const { setLoading } = useContext<ILoading>(LoadingContext);
+const Leaderboards: Component = () => {
+  const [data, setData] = createSignal<LeaderboardData[]>([]);
 
-  useEffect(() => {
-    setLoading(true);
+  createEffect(() => {
+    LoadingContext.loading = true;
     UserService.getAllStats()
     .then((response:any) => {
-      setLoading(false);
+      LoadingContext.loading = false;
       setData(response.data
         .sort((row1:GetStat, row2:GetStat) => row2.highScore - row1.highScore )
         .map((row:GetStat) => {
@@ -29,20 +26,25 @@ const Leaderboards: React.FC = () => {
       }))
     })
     .catch((err:any) => {
-      setLoading(false);
-      toast.error(err.response.data.message);
+      LoadingContext.loading = false;
+      // createToast({
+      //   title: err.response.data.message,
+      //   duration: 2000,
+      //   placement: "top-end"
+      // })
     })
      // eslint-disable-next-line
-  }, [useEffect]);
+  }, [createEffect]);
     
   let test = {
     columns: [{label: "User", field: "user"}, {label: "High Score", field: "highScore", sort: "desc"}],
     rows: data,
   }
   return (
-    <div className="d-flex justify-content-center align-items-center mt-1" style={{height: "100%"}}>
-			<div className="row justify-content-center w-75">
-        <MDBDataTableV5 hover scrollY searching={false} sortable order={['High Score', 'desc']} paging={false} data={test as any} />
+    <div class="d-flex justify-content-center align-items-center mt-1" style={{height: "100%"}}>
+			<div class="row justify-content-center w-75">
+        {/* todo replace with table */}
+        {/* <MDBDataTableV5 hover scrollY searching={false} sortable order={['High Score', 'desc']} paging={false} data={test as any} /> */}
       </div>
     </div>
   );
